- Add NerMetrics for NER DL and CRF - Refactor eval packages names
