+++ eval/src/main/scala/com/johnsnowlabs/nlp/eval/ner/NerDLEvaluation.scala
@@ -1,11 +1,11 @@
-package com.johnsnowlabs.nlp.eval
+package com.johnsnowlabs.nlp.eval.ner
 
 import com.johnsnowlabs.nlp.annotator.{NerDLApproach, NerDLModel, WordEmbeddings, WordEmbeddingsModel}
 import com.johnsnowlabs.nlp.eval.util.{GoldTokenizer, LoggingData}
 import com.johnsnowlabs.nlp.training.CoNLL
 import org.apache.spark.mllib.evaluation.MulticlassMetrics
-import org.apache.spark.sql.{Dataset, SparkSession}
 import org.apache.spark.sql.functions._
+import org.apache.spark.sql.{Dataset, SparkSession}
 
 import scala.collection.mutable
 
@@ -40,9 +40,9 @@ class NerDLEvaluation(sparkSession: SparkSession, testFile: String, tagLevel: St
     val predictionLabelsRDD = evaluationDataSet.select("predictionIndex", "labelIndex")
       .map(r => (r.getDouble(0), r.getDouble(1)))
     val metrics = new MulticlassMetrics(predictionLabelsRDD.rdd)
-    computeAccuracy(metrics)
-    computeAccuracyByEntity(metrics, entityLabels)
-    computeMicroAverage(metrics)
+    NerMetrics.computeAccuracy(metrics, loggingData)
+    NerMetrics.computeAccuracyByEntity(metrics, entityLabels, loggingData)
+    NerMetrics.computeMicroAverage(metrics,loggingData)
   }
 
   private def getEntityLabels(nerEvalDLConfiguration: NerEvalDLConfiguration, tagLevel: String): List[String] = {
@@ -58,36 +58,6 @@ class NerDLEvaluation(sparkSession: SparkSession, testFile: String, tagLevel: St
     }
 
     entities.toList.distinct
-  }
-
-  private def getEvaluationDataSet(nerEvalDLConfiguration: NerEvalDLConfiguration, entityLabels: List[String],
-                                   tagLevel: String): Dataset[_] = {
-
-    val joinedDataSet = getJoinedDataSet(nerEvalDLConfiguration)
-    val labelAndPrediction: Seq[(String, String)] = joinedDataSet.select("testTags", "predictedTags").rdd.map { row =>
-      val labelColumn: Seq[String] = row.get(0).asInstanceOf[mutable.WrappedArray[String]]
-      val predictionColumn: Seq[String] = row.get(1).asInstanceOf[mutable.WrappedArray[String]]
-      if (tagLevel == "IOB") {
-        (labelColumn.toList, predictionColumn.toList)
-      } else {
-        val groupLabelColumn = labelColumn.map(element => element.replace("I-", "")
-          .replace("B-", ""))
-        val groupPredictionColumn = predictionColumn.map(element => element.replace("I-", "")
-          .replace("B-", ""))
-        (groupLabelColumn.toList, groupPredictionColumn.toList)
-      }
-
-    }.collect().flatMap(row => row._1 zip row._2)
-
-    labelAndPrediction.toDF("label", "prediction")
-      .withColumn("labelIndex", getLabelIndex(entityLabels)(col("label")))
-      .withColumn("predictionIndex", getLabelIndex(entityLabels)(col("prediction")))
-      .filter(col("label") =!= 'O')
-  }
-
-  private def getLabelIndex(labels: List[String]) = udf { label: String =>
-    val index = labels.indexOf(label)
-    index.toDouble
   }
 
   private def getJoinedDataSet(nerEvalDLConfiguration: NerEvalDLConfiguration): Dataset[_] = {
@@ -147,50 +117,34 @@ class NerDLEvaluation(sparkSession: SparkSession, testFile: String, tagLevel: St
     )
   }
 
-  private def computeAccuracy(metrics: MulticlassMetrics): Unit = {
-    val accuracy = (metrics.accuracy * 1000).round / 1000.toDouble
-    val weightedPrecision = (metrics.weightedPrecision * 1000).round / 1000.toDouble
-    val weightedRecall = (metrics.weightedRecall * 1000).round / 1000.toDouble
-    val weightedFMeasure = (metrics.weightedFMeasure * 1000).round / 1000.toDouble
-    val weightedFalsePositiveRate = (metrics.weightedFalsePositiveRate * 1000).round / 1000.toDouble
-    loggingData.logMetric("Accuracy", accuracy)
-    loggingData.logMetric("Weighted Precision", weightedPrecision)
-    loggingData.logMetric("Weighted Recall", weightedRecall)
-    loggingData.logMetric("Weighted F1 Score", weightedFMeasure)
-    loggingData.logMetric("Weighted False Positive Rate", weightedFalsePositiveRate)
+  private def getEvaluationDataSet(nerEvalDLConfiguration: NerEvalDLConfiguration, entityLabels: List[String],
+                                   tagLevel: String): Dataset[_] = {
+
+    val joinedDataSet = getJoinedDataSet(nerEvalDLConfiguration)
+    val labelAndPrediction: Seq[(String, String)] = joinedDataSet.select("testTags", "predictedTags").rdd.map { row =>
+      val labelColumn: Seq[String] = row.get(0).asInstanceOf[mutable.WrappedArray[String]]
+      val predictionColumn: Seq[String] = row.get(1).asInstanceOf[mutable.WrappedArray[String]]
+      if (tagLevel == "IOB") {
+        (labelColumn.toList, predictionColumn.toList)
+      } else {
+        val groupLabelColumn = labelColumn.map(element => element.replace("I-", "")
+          .replace("B-", ""))
+        val groupPredictionColumn = predictionColumn.map(element => element.replace("I-", "")
+          .replace("B-", ""))
+        (groupLabelColumn.toList, groupPredictionColumn.toList)
+      }
+
+    }.collect().flatMap(row => row._1 zip row._2)
+
+    labelAndPrediction.toDF("label", "prediction")
+      .withColumn("labelIndex", getLabelIndex(entityLabels)(col("label")))
+      .withColumn("predictionIndex", getLabelIndex(entityLabels)(col("prediction")))
+      .filter(col("label") =!= 'O')
   }
 
-  private def computeAccuracyByEntity(metrics: MulticlassMetrics, labels: List[String]): Unit = {
-    val predictedLabels = metrics.labels
-    predictedLabels.foreach { predictedLabel =>
-      val entity = labels(predictedLabel.toInt)
-      val precision = (metrics.precision(predictedLabel) * 1000).round / 1000.toDouble
-      val recall = (metrics.recall(predictedLabel) * 1000).round / 1000.toDouble
-      val f1Score = (metrics.fMeasure(predictedLabel) * 1000).round / 1000.toDouble
-      val falsePositiveRate = (metrics.falsePositiveRate(predictedLabel) * 1000).round / 1000.toDouble
-      loggingData.logMetric(entity + " Precision", precision)
-      loggingData.logMetric(entity + " Recall", recall)
-      loggingData.logMetric(entity + " F1-Score", f1Score)
-      loggingData.logMetric(entity + " FPR", falsePositiveRate)
-    }
-  }
-
-  private def computeMicroAverage(metrics: MulticlassMetrics): Unit = {
-    var totalP = 0.0
-    var totalR = 0.0
-    var totalClassNum = 0
-
-    val labels = metrics.labels
-
-    labels.foreach { label =>
-      totalClassNum = totalClassNum + 1
-      totalP = totalP + metrics.precision(label)
-      totalR = totalR + metrics.recall(label)
-    }
-    totalP = totalP/totalClassNum
-    totalR = totalR/totalClassNum
-    val microAverage = 2 * ((totalP*totalR) / (totalP+totalR))
-    loggingData.logMetric("Micro-average F1-Score", (microAverage * 1000).round / 1000.toDouble)
+  private def getLabelIndex(labels: List[String]) = udf { label: String =>
+    val index = labels.indexOf(label)
+    index.toDouble
   }
 
 }
