+++ eval/src/main/scala/com/johnsnowlabs/nlp/eval/spell/SymSpellEvaluation.scala
@@ -1,4 +1,4 @@
-package com.johnsnowlabs.nlp.eval
+package com.johnsnowlabs.nlp.eval.spell
 
 import com.johnsnowlabs.nlp.annotator._
 import com.johnsnowlabs.nlp.annotators._
